add a filter class, authentication exception, and a cors filter
filter class that extends OncePerRequestFilter authenticates all
other requests. checks the Authorization header and validates the token
and fetches the user to setAuthentication in Security context holder.
then the filter is added in filterChain as addFilterBefore.

authentication exception is implemented by AuthenticationEntryPoint
and sets the response to 403 and prints out a Error message.
it's added into the filterChain as exceptionHandling.

CORS filter is added into the SecurityConfig and defines allowed origins,
methods and headers to be used. This is needed for a front-end sending
requests from a different origin. CORS filter intercepts the requests and
add an appropriate header to the requests.